+++ MasterPlanner/src/block/BlockState.java
@@ -2,24 +2,24 @@ package block;
 
 public class BlockState {
 	
-	final public static int Unknown = 0;
-	final public static int Located = 1;
-	final public static int Missing = 2;
-	final public static int Moving = 3;
-	final public static int Delivered = 4;
+	final public static byte Unknown = 0;
+	final public static byte Located = 1;
+	final public static byte Missing = 2;
+	final public static byte Moving = 3;
+	final public static byte Delivered = 4;
 	
 	final public static String[] blockStates = {
 		"Unknown", "Located", "Missing", "Moving", "Delivered"
 	};
 	
-	public static boolean isValid(int state) {
+	public static boolean isValid(byte state) {
 		return state >= 0 && state < blockStates.length;
 	}
 	
-	public static int parseFrom(String data) {
+	public static byte parseFrom(String data) {
 		if(data == null) { return -1; }
 		String temp = data.trim();
-		for(int i=0;i<blockStates.length;i++) {
+		for(byte i=0;i<blockStates.length;i++) {
 			if(temp.equalsIgnoreCase(blockStates[i])) {
 				return i;
 			}
@@ -27,7 +27,7 @@ public class BlockState {
 		return -1;
 	}
 	
-	public static String toString(int blockState) {
+	public static String toString(byte blockState) {
 		return !isValid(blockState) ? "Invalid" : blockStates[blockState];
 	}
