+++ MasterPlanner/src/pot/PotState.java
@@ -2,22 +2,22 @@ package pot;
 
 public class PotState {
 	
-	final public static int Origin = 0;
-	final public static int Moving = 1;
-	final public static int Delivered = 2;
+	final public static byte Origin = 0;
+	final public static byte Moving = 1;
+	final public static byte Delivered = 2;
 	
 	final public static String[] potStates = {
 		"Origin", "Moving", "Delivered"
 	};
 	
-	public static boolean isValid(int state) {
+	public static boolean isValid(byte state) {
 		return state >= 0 && state < potStates.length;
 	}
 	
-	public static int parseFrom(String data) {
+	public static byte parseFrom(String data) {
 		if(data == null) { return -1; }
 		String temp = data.trim();
-		for(int i=0;i<potStates.length;i++) {
+		for(byte i=0;i<potStates.length;i++) {
 			if(temp.equalsIgnoreCase(potStates[i])) {
 				return i;
 			}
@@ -25,7 +25,7 @@ public class PotState {
 		return -1;
 	}
 	
-	public static String toString(int potState) {
+	public static String toString(byte potState) {
 		return !isValid(potState) ? "Invalid" : potStates[potState];
 	}
